Split resolver metadata validation descriptor tests
